move constructTypingString to util
