+++ src/components/MessageList/utils/constructTypingString.js
@@ -1,0 +1,38 @@
+export const constructTypingString = ({ client, t, typingUsers }) => {
+  const typingKeys = Object.keys(typingUsers);
+  const nonSelfUsers = [];
+  typingKeys.forEach((item, i) => {
+    if (client.user.id === typingUsers[typingKeys[i]].user.id) {
+      return;
+    }
+    nonSelfUsers.push(
+      typingUsers[typingKeys[i]].user.name ||
+        typingUsers[typingKeys[i]].user.id,
+    );
+  });
+
+  let outStr = '';
+  if (nonSelfUsers.length === 1) {
+    outStr = t('{{ user }} is typing...', { user: nonSelfUsers[0] });
+  } else if (nonSelfUsers.length === 2) {
+    /**
+     * Joins the two names without commas
+     * example: "bob and sam"
+     */
+    outStr = t('{{ firstUser }} and {{ secondUser }} are typing...', {
+      firstUser: nonSelfUsers[0],
+      secondUser: nonSelfUsers[1],
+    });
+  } else if (nonSelfUsers.length > 2) {
+    /**
+     * Joins all names with commas, the final one gets ", and" (oxford comma!)
+     * example: "bob, joe, and sam"
+     */
+    outStr = t('{{ commaSeparatedUsers }} and {{ lastUser }} are typing...', {
+      commaSeparatedUsers: nonSelfUsers.slice(0, -1).join(', '),
+      lastUser: nonSelfUsers.slice(-1),
+    });
+  }
+
+  return outStr;
+};
