+++ src/components/MessageList/utils/index.js
@@ -1,3 +1,4 @@
+export { constructTypingString } from './constructTypingString';
 export { getGroupStyles } from './getGroupStyles';
 export { getLastReceivedId } from './getLastReceivedId';
 export { getReadStates } from './getReadStates';
